Delete multiple roles.
Former-commit-id: 325b47de9e39709c9f4863a41025d0b1742b6586

+++ PLATFORM/Modules/Core/VirtoCommerce.Core.Web/Controllers/Api/SecurityController.cs
@@ -122,12 +122,24 @@ namespace VirtoCommerce.CoreModule.Web.Controllers.Api
             return Ok(result);
         }
 
+        /// <summary>
+        /// DELETE: api/security/roles?id=1&amp;id=2
+        /// </summary>
+        /// <param name="roleIds"></param>
+        /// <returns></returns>
         [HttpDelete]
-        [Route("roles/{roleId}")]
+        [Route("roles")]
         [ResponseType(typeof(RoleDescriptor))]
-        public IHttpActionResult DeleteRole(string roleId)
+        public IHttpActionResult DeleteRoles([FromUri(Name = "id")] string[] roleIds)
         {
-            _roleService.DeleteRole(roleId);
+            if (roleIds != null)
+            {
+                foreach (var roleId in roleIds)
+                {
+                    _roleService.DeleteRole(roleId);
+                }
+            }
+
             return Ok();
         }
